Fixed sign compare warnings by changing or adding casts

+++ volume_io/Volumes/output_mnc.c
@@ -1449,7 +1449,7 @@ static  VIO_Status  output_the_volume(
           file->n_slab_dims++;  /* integral number of complete dimensions */
         } else {
           count[d] = MIN( volume_count[to_volume_index[d]], 
-                          (size_t)( MI_MAX_VAR_BUFFER_SIZE / ( slab_size * unit_size ) ) );
+                          (long)( MI_MAX_VAR_BUFFER_SIZE / ( slab_size * unit_size ) ) );
           n_steps *= (int)( ( volume_count[to_volume_index[d]] + count[d] - 1 ) / count[d] );
         }
         slab_size *= count[d];
